Rename fixture file for swapMinAmounts tests
